Refactor on the SDK
