+++ packages/gitignoreio-sdk/src/__tests__/data.ts
@@ -1,5 +1,5 @@
 import { mock } from 'vitest-mock-extended';
 
-import type { HttpClient } from '../domain';
+import type { HttpClient } from '../domain/http-client';
 
 export const makeMockHttpClient = () => mock<HttpClient>();
