+++ packages/gitignoreio-sdk/src/adapters/sdk.ts
@@ -1,6 +1,8 @@
-import type { GitIgnoreInput, GitIgnoreIoSDK, HttpClient } from './domain';
+import type { GitIgnoreIoSDK } from '../domain';
+import type { Technologies } from '../domain/technology';
 
-import { DefaultHttpClient } from './adapters/fetch/client';
+import { DefaultHttpClient } from '../adapters/fetch/client';
+import { type HttpClient } from '../domain/http-client';
 
 /**
  * GitIgnore SDK for generating .gitignore files
@@ -25,12 +27,20 @@ export class GitIgnoreSDK implements GitIgnoreIoSDK {
    * Generates .gitignore content for the given technologies by calling the {@link https://docs.gitignore.io/use/api} API.
    *
    * @param technologies Array of technology identifiers to include in the .gitignore.
-   * @returns A ResultAsync, either containing the generated .gitignore content or an error.
+   * @returns A Promise that resolves to the generated .gitignore content or rejects with an error.
    */
-  generate(technologies: GitIgnoreInput) {
+  async generate(technologies: Technologies) {
     const url = new URL(
       `${this.baseUrl}/${encodeURIComponent(technologies.join(','))}`,
     );
-    return this.httpClient.get(url).map((content) => ({ content }));
+    const result = await this.httpClient.get(url);
+    return result
+      .map((content) => ({ content }))
+      .match(
+        (value) => value,
+        () => {
+          throw new Error('There was an error fetching data from gitignore.io');
+        },
+      );
   }
 }
