+++ packages/gitignoreio-sdk/src/domain/http-client.ts
@@ -1,0 +1,8 @@
+import type { ResultAsync } from 'neverthrow';
+
+/**
+ * Interface for HTTP client dependency.
+ */
+export interface HttpClient {
+  get(url: URL): ResultAsync<string, Error>;
+}
