+++ packages/gitignoreio-sdk/src/domain/index.ts
@@ -1,15 +1,7 @@
-import type { ResultAsync } from 'neverthrow';
-
-import type { GitIgnoreElement } from './gitignore-element';
-
-/**
- * Input for generating gitignore content.
- * A non-empty array of technology names.
- */
-export type GitIgnoreInput = [GitIgnoreElement, ...GitIgnoreElement[]];
+import type { Technologies } from './technology';
 
 export interface GitIgnoreIoSDK {
-  generate(technologies: GitIgnoreInput): ResultAsync<GitIgnoreResult, Error>;
+  generate(technologies: Technologies): Promise<GitIgnoreResult>;
 }
 
 /**
@@ -17,11 +9,4 @@ export interface GitIgnoreIoSDK {
  */
 export interface GitIgnoreResult {
   content: string;
-}
-
-/**
- * Interface for HTTP client dependency.
- */
-export interface HttpClient {
-  get(url: URL): ResultAsync<string, Error>;
 }
